+++ lwlog/src/details/os/os_impl.h
@@ -3,14 +3,14 @@
 namespace lwlog::details::os
 {
 	template<typename ThreadIdPolicy>
-	std::size_t get_thread_id()
+	std::uint64_t get_thread_id()
     {
-		std::size_t thread_id{};
+		std::uint64_t thread_id{};
 
 		#if defined(_WIN32)
-			thread_id = static_cast<std::size_t>(::GetCurrentThreadId());
+			thread_id = static_cast<std::uint64_t>(::GetCurrentThreadId());
 		#elif defined(__linux__)
-			thread_id = static_cast<std::size_t>(::syscall(SYS_gettid));
+			thread_id = static_cast<std::uint64_t>(::syscall(SYS_gettid));
 		#elif defined(__APPLE__)
 			::pthread_threadid_np(NULL, &thread_id);
 		#else
@@ -21,35 +21,35 @@ namespace lwlog::details::os
 	}
 
 	template<typename ProcessIdPolicy>
-	std::size_t get_process_id()
+	std::uint64_t get_process_id()
 	{
 		#if defined(_WIN32)
-			return static_cast<std::size_t>(::GetCurrentProcessId());
+			return static_cast<std::uint64_t>(::GetCurrentProcessId());
 		#elif defined(__linux__) || defined(__APPLE__)
-			return static_cast<std::size_t>(::getpid());
+			return static_cast<std::uint64_t>(::getpid());
 		#endif
 	}
 
 	template<> 
-	std::size_t get_thread_id<disable_thread_id>() 
+	std::uint64_t get_thread_id<disable_thread_id>()
 	{ 
 		return {}; 
 	}
 
 	template<> 
-	std::size_t get_process_id<disable_process_id>() 
+	std::uint64_t get_process_id<disable_process_id>()
 	{ 
 		return {}; 
 	}
 
 	template<typename ThreadIdPolicy, typename ProcessIdPolicy>
-	std::size_t execution_context<ThreadIdPolicy, ProcessIdPolicy>::thread_id() const
+	std::uint64_t execution_context<ThreadIdPolicy, ProcessIdPolicy>::thread_id() const
 	{ 
 		return get_thread_id<ThreadIdPolicy>();
 	}
 
 	template<typename ThreadIdPolicy, typename ProcessIdPolicy>
-	std::size_t execution_context<ThreadIdPolicy, ProcessIdPolicy>::process_id() const
+	std::uint64_t execution_context<ThreadIdPolicy, ProcessIdPolicy>::process_id() const
 	{ 
 		return get_process_id<ProcessIdPolicy>();
 	}
